Refactor instance creation methods
